backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "artemarket-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "python-jose[cryptography]>=3.3",
    "pydantic-settings>=2.0",
    "apscheduler>=3.10",
    "psycopg2-binary>=2.9",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx"]
